+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 - Changed isBool to isBoolean, enforce.bool to enforce.boolean, and method.bool to method.boolean
 - Changed isFunc to isFunction, enforce.func to enforce.function, and method.func to method.function
+- Changed isInt to isInteger, enforce.int to enforce.integer, and method.int to method.integer
 
 ## 0.2.0 - 2019-01-18
 ### Added